+++ packages/widgetbook/lib/src/fields/boolean_field.dart
@@ -12,6 +12,7 @@ class BooleanField extends Field<bool> {
     super.initialValue = true,
     @Deprecated('Fields should not be aware of their context') super.onChanged,
   }) : super(
+         defaultValue: true,
          type: FieldType.boolean,
          codec: FieldCodec(
            toParam: (value) => value.toString(),