+++ packages/widgetbook/lib/src/fields/string_field.dart
@@ -13,6 +13,7 @@ class StringField extends Field<String> {
     this.maxLines,
     @Deprecated('Fields should not be aware of their context') super.onChanged,
   }) : super(
+         defaultValue: '',
          type: FieldType.string,
          codec: FieldCodec(
            toParam: (value) => value,